Cache anki collection in Google Cloud Storage
The initial sync of the collection is quite slow (up to 3 minutes). Avoid
running it every time by caching the collection. Since we are dealing
with many small media files, putting everything into a tar archive reduces the
amount of requests to Google Cloud Storage and speeds things up a bit.
Adding gzip on top actually worsened the runtime, so we don't use it.

After these changes the function runs between 10 to 15 seconds including
downloading the collection, processing it, and uploading it.

Signed-off-by: David Sauerwein <david@sauerwein.xyz>